requirements.txt
# Core Dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Attribution Methods
captum>=0.6.0

# Data Handling
pandas>=1.5.0
scikit-learn>=1.0.0
Pillow>=8.0.0
scipy>=1.9.0

# Visualization
seaborn>=0.11.0
opencv-python>=4.5.0

# Optional: Advanced Features
tensorboard>=2.8.0  # For logging and monitoring
transformers>=4.20.0  # For Hugging Face ViT models

# Development (Optional)
jupyter>=1.0.0  # For notebook analysis
ipywidgets>=7.7.0  # For interactive widgets